Make Image Registry host optional and some README changes
If you are using dockerhub, you don't the image registry host
is the same as your username. It's kind of confusing to the
end user. We can make it optional and default it to the username
if not specified (i.e assumer docker hub).

Some other changes include changing the domain to example.com and
formatting the README's QuickStart.