Removes specifying hostname and renames registry params
The patch removes the need to specify an hostname during dispatch
installation and hence adding the hostname to your local
/etc/hosts. This is done by removing name based virtual routing
in ingress. But the user has to specify at least the IP address
of the dispatch host i.e the node address if using NodePort or the
LB IP if using loadbalancer.

The patch also renames image registry's host parameter to 'name'.
This is because the term host is misleading for a registry which
may also contain the username or repo details as part of
the it. e.g gcr.io/<PROJECT_ID>