Update README to mention PV requirement
Dispatch requires PV dynamic provisioning support in order to
deploy the rabbitmq helm chart. We could also specify existing
PVC in the chart but it will be tedious to get this from user
during installation.

This is a not a major constraint since for a single-node K8S
installations, user could use hostpath provisioner (similar
to minikube environments).