Check Helm release and add insecure registry config
Add a validation to check if there are failed helm releases and if
so ask the user to purge them. Added a flag --ignore-helm-check
if the user knows what she/he is doing and wants to continue
with the install.

Added an insecure option to ImageRegistry since all registries
were treated as insecure. Specifying a registry like
quay.io/<username> will fail dind pod since it is not
a valid hostname or CIDR range for --insecure-registry option.