Support for chart options and single ns uninstall
Currently, it's not possible to customize any helm chart value
using the dispatch install command. Only a limited set of opts
are exposed in the install configuration. But for advanced
installations, the deployer may want to specify additional helm
chart values. e.g setting resource limits at
a granular level or disabling rbac on specific charts etc.
It will be tedious and redundant to re-expose all chart
values as dispatch opts.

Changes in this patch :-
1. Add Opts field to Chart for getting additional chart
   configuration.
2. Flag to support specifying tiller-namespace if not in
   kube-system.
3. Uninstall support if installed in single-namespace mode.