Support for default resource requirements in functions
This patch adds support to specify a default for the resource requirements
of a function. This allows a dispatch operator to customize the default
resource allocated to a function and also set limits.

Both OpenFaaS and Riff support this feature. We can extend this
implementation to specify per-function resource requests & limits.

Included in this patch :
1. Bump faas-netes to 0.4.2 where resource requests & limits are
   supported.
2. Add Chat values for specifying resource requirements
3. Enhance the drivers to pass resource requirements