Fix for getting deployment objects in K8S 1.7
Currently, function manager can't get deployments in K8S 1.7.
This will result in failed function creations even though
it was successful on OpenFaaS and a deployment object was
created on K8S. This is because we use apps/v1beta2 apigroup
which was introduced only in K8S 1.8.

Let's switch to apps/v1beta1 which will still be
autoconverted post 1.9. Once we remove support for 1.7 clusters,
we can switch to apps/v1beta2 or apps/v1 when we move to 1.9.

Reference:- https://kubernetes.io/docs/reference/workloads-18-19/