Add Kafka backend for events and make it default (#351)
**Changes**:
* adds new event transport based on Kafka Broker
* kafka is now the default transport regardless of FaaS used. RabbitMQ
can still be enabled via config flag.
* Kafka is now deployed to dispatch namespace instead of separate one,
and an external deployment can be used (before it was hardcoded to use
the built-in one).

**Pending Issues**
Kafka does not support wildcard topic matching. Because of that,
subscriptions are now 1:1 (source and event types must be set without
`*`). I will create separate issue to discuss wildcard subscriptions (as
it's related to multi tenancy, performance, and other aspects)