edit Dispatch license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit updates the LICENSE file so that it contains only the full
text of the Apache license and references VMWare, Inc as the copyright
holder. The info that pertains to 3rd-party software licensing has now
been moved to a new DEPENDENCIES file. Finally, the README file has a
new "Licensing" section to explain to readers the purposes of the
LICENSE and DEPENDENCIES files.

Collectively, these changes allow Licensee to successfully identify the
license type of Dispatch as Apache.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>