Use command string in generate script to ensure consistency (#442)
In previous patch the generate-models.sh had an incorrect setting for
one of the variables (WORK_DIR). It wasn't caught in CI, because
CI used different command (direct swagger call instead of docker).
This patch fixes the WORK_DIR for generate-models, and also extracts
the command to a shared so variable, so that CI and local execute the
same one.