Add tracing context to function context
Function manager uses dedicated Context type, not compatible
with context.Context. This patch adds tracing context as one of the keys
in function context, and adds a middleware to remove it right before
function execution. The tracing context is used to trace middlewares and
correlate requests to other services (secret and service managers).