Fix datatype for entity store regarding timestamp
The previous data type only stored time, not date making the created
and modified times useless.  They are intended to be timetstamps and
the CLI expects that.

This change is a breaking change as there is no data migration.

Also lengthens the retries in e2e where we were seeing some failures.

Fixes: #485