Support Tenancy in Dispatch (#510)
* Introduce Tenancy Support in Dispatch

* CLI changes

* Swagger changes

* Remove auth requirements for version

* Add ListGlobal method and other refactor for entity store

* Add Global Policy, fix CI and CLI

In order to denote a policy that applies to all orgs, a global
flag is added to the model. When specified, the policy applies
to all orgs (in other words god mode). Without this it's not
possible to give a single user/account access to all orgs.

Also, fixed issues with service account in multi-tenant mode.

* Rebase after Identity Manager CLI client changes