Make the event transport compatible with recent multi-org changes (#531)
* Transport interface now expects organization for both Publish and
Subscribe methods. For both Kafka and RabbitMQ, organization is
prepended to the topic. This has certain implications on organization
name/allowed characters. If this becomes a problem, we should consider
using some other Organization ID.
* RabbitMQ now uses single Exchange. Previous approach was to declare
separate Exchange for each organization, but this would require event
manager to listen on changes in Organization entities. For now,
a simpler approach of using single exchange and prepending the org to
the topic is used.
* Cosmetic improvements in error handling & wording