Updates e2e to run against Dispatch solo
Adds a new pipeline that builds the dispatch-server and dispatch-cli
binaries, and then runs the e2e tests against a dispatch-server running
locally. Currently disables event, services, login and organization test
suites, and relies on the http endpoint for API tests. Once we have a
way to add certs that will need to be modified.