Build lightweight VMware guest tools alternative
Use the govmomi toolbox library to build a lightweight VMware guest
tools implementation. This will allow for graceful power operations
to be initiated via the hypervisor, reporting of DHCP-assigned IP,
and other quality-of-life improvements, without depending on the
proprietary guest tools product.

The implementation is based on the [example code][1] and some of the
modifications from the [vic-product implementation][2].

[1]:https://github.com/vmware/govmomi/blob/master/toolbox/toolbox/main.go
[2]:https://github.com/vmware/vic-product/blob/master/installer/toolbox/main.go