Build a virtual appliance for Dispatch  (#680)
* Build lightweight VMware guest tools alternative

Use the govmomi toolbox library to build a lightweight VMware guest
tools implementation. This will allow for graceful power operations
to be initiated via the hypervisor, reporting of DHCP-assigned IP,
and other quality-of-life improvements, without depending on the
proprietary guest tools product.

The implementation is based on the [example code][1] and some of the
modifications from the [vic-product implementation][2].

[1]:https://github.com/vmware/govmomi/blob/master/toolbox/toolbox/main.go
[2]:https://github.com/vmware/vic-product/blob/master/installer/toolbox/main.go

* Add tool for managing guestinfo values

The rpctool allows for easy in-guest management of guestinfo values.

The implementation is based on the [vic-product implementation][1].

[1]:https://github.com/vmware/vic-product/blob/master/installer/ovatools/rpctool/main.go

* Add tool for accessing OVF environment values

The ovfenv utility allows for easy in-guest access to OVF environment
values, which will be used to set the initial root password and to
provide the server certificate and key.

The implementation is based on the [vic-product implementation][1].

[1]:https://github.com/vmware/vic-product/blob/master/installer/ovatools/ovfenv/main.go

* Build a virtual appliance for Dispatch

Introduces an ova make target, which builds an OVA for dispatch-sever
using the scripts under scripts/ova.

The OVA has two disks. The first is treated as the root filesystem and
the second is used for application config files, data, and logs.

The OVA is Photon-based, and includes a variety of dependencies, the
dispatch-server binary, and additional utility binaries: toolbox (a
lightweight VMware guest tools implementation), ovfenv (for accessing
OVF environment properties), and rpctool (for managing guestinfo).

Systemd units handle configuration of storage and networking.

Basic configuration information, such as the initial root password, is
passed to the OVA via environment properties.

The OVA is built using a container. The build.sh script can be called
from the container (e.g., during CI) with the ova-ci argument or will
launch the container itself with the ova-dev argument.

The OVA build process creates raw disk images, installs the base OS,
and then configures the OS in application-specific ways. The disk
images are then converted into VMDKs. Next, placeholder information is
replaced in an OVF file. Lastly, the files are compressed into an OVA.