Fix image creation/deletion issue
There is a race condition with images where image might be in deleting
state, causing list API to not include it in the result, but also
not allowing to create a new one with the same name. This patch
attempts to fix it by always returning all images in the list, regardless of their
state, until the actual record is deleted from the database.

Also makes e2e number of retries and wait times configurable via global
variable (can still be specified in the retry function itself)